fix skip_memcheck
compare the value of skip_memcheck instead of testing if it is not empty.
skip_memcheck must be quoted, otherwise bash will treat it like a command
and run it (will execute the builtin "true" or "false).

VAL=value ; if [ $VAL ] ; then echo "foo" ; else echo "bar" ; fi

will print foo regardless if value is true or false, because this
is equivalent to [ -n $VAL ] (verifies that length of output of command
$VAL is nonzero).